LayoutTests/webanimations/keyframe-effect-interface-timing-duration.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that KeyframeEffect objects expose a duration through its timing property.");

debug("Check the default state of the timing and duration properties.");
const keyframeEffect = new KeyframeEffect(document.body);
shouldBeDefined("keyframeEffect.timing");
shouldBeDefined("keyframeEffect.timing.duration");
shouldBe("keyframeEffect.timing.duration", "0");

debug("");
debug("Check that we can update the timing duration.");
keyframeEffect.timing.duration = 2;
shouldBe("keyframeEffect.timing.duration", "2");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
