LayoutTests/webanimations/keyframe-effect-interface-timing-duration-expected.txt

Check that KeyframeEffect objects expose a duration through its timing property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check the default state of the timing and duration properties.
PASS keyframeEffect.timing is defined.
PASS keyframeEffect.timing.duration is defined.
PASS keyframeEffect.timing.duration is 0

Check that we can update the timing duration.
PASS keyframeEffect.timing.duration is 2

PASS successfullyParsed is true

TEST COMPLETE
